Render image with screen circumscribed
